% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mintde.R
\name{apply_deseq2}
\alias{apply_deseq2}
\title{Apply DeSeq2}
\usage{
apply_deseq2(dat, hours = c(0, 3, 6, 12, 24, 48, 72))
}
\arguments{
\item{dat}{a matrix of LFC, row for the object, column for the timepoint.}

\item{hours}{an array of timepoints}
}
\value{
list
}
\description{
Apply DeSeq2 to different timepoints against baseline.
}
