% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mintde.R
\name{Edgington_method}
\alias{Edgington_method}
\title{Gene Selection based on Edgington's Method}
\usage{
Edgington_method(topk, p1, p2, p3, Genes)
}
\arguments{
\item{topk}{an array of topk genes to choose}

\item{p1}{array of p-values for the strength of DE in the first modality}

\item{p2}{array of p-values for the strength of DE in the second modality}

\item{p3}{array of p-values for correlation over time across modalities}

\item{Genes}{array of gene names}
}
\value{
gene_id gives the array of selected genes, idx gives the corresponding index
}
\description{
The core function that select genes differentially expressed (DE) and correlated across modalitiesc based on givens p-values using Edgington's Method (sum of p-values).
}
\examples{
Edgington_method(topk=c(10,50),
p1=runif(1000,min=0,max=0.05),
p2=runif(1000,min=0,max=0.05),
p3=runif(1000,min=0,max=0.05),
Genes=paste0('gene',1:1000))
}
