% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mintde.R
\name{select_genes}
\alias{select_genes}
\title{Select Genes}
\usage{
select_genes(param, method, p1, p2, p3 = NULL, Genes = NULL)
}
\arguments{
\item{param}{parameter associated with each method. For Edgington (sum of p-values) and Fisher (sum of log p-values), an array of topk genes to choose; For commonDE (only thresholding p1 and p2), an array of threshold of p1 and p2}

\item{method}{string of the method to use, "E" for Edgington (sum of p-values), "F" for Fisher (sum of log p-values), and "V" for commonDE/Venn diagram method (only thresholding p1 and p2)}

\item{p1}{array of p-values for the strength of DE in the first modality}

\item{p2}{array of p-values for the strength of DE in the second modality}

\item{p3}{array of p-values for correlation over time across modalities}

\item{Genes}{array of gene names}
}
\value{
an array of selected genes
}
\description{
The core function that select genes differentially expressed (DE) and correlated across modalitiescbased on givens p-values.
}
