% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mintde.R
\name{Venn_method}
\alias{Venn_method}
\title{commonDE/Venn Diagram Method}
\usage{
Venn_method(cutoff, p1, p2, Genes)
}
\arguments{
\item{cutoff}{an array of cutoff (=10^(cutoff)) of p-values}

\item{p1}{array of p-values for the strength of DE in the first modality}

\item{p2}{array of p-values for the strength of DE in the second modality}

\item{Genes}{array of gene names}
}
\value{
an array of selected genes
}
\description{
commonDE/Venn diagram method (only thresholding p-values from two modalities)
}
\examples{
Venn_method(cutoff=c(-2,-3),
p1=runif(1000,min=0,max=0.05),
p2=runif(1000,min=0,max=0.05),
Genes=paste0('gene',1:1000))
}
